fix tests

+++ core/src/test/java/org/elasticsearch/search/aggregations/bucket/DateScriptMocksPlugin.java
@@ -45,7 +45,8 @@ public class DateScriptMocksPlugin extends MockScriptPlugin {
             String fieldname = (String) params.get("fieldname");
             return docLookup.get(fieldname);
         });
-        scripts.put(PLUS_ONE_MONTH, params -> new DateTime((long) params.get("_value"), DateTimeZone.UTC).plusMonths(1).getMillis());
+        scripts.put(PLUS_ONE_MONTH, params ->
+            new DateTime(Double.valueOf((double) params.get("_value")).longValue(), DateTimeZone.UTC).plusMonths(1).getMillis());
         return scripts;
     }
 }
